attendance-master/requirements.txt
asgiref                 3.4.1
click                   8.0.3
colorama                0.4.4
Django                  3.2.3
dlib                    19.22.1
face-recognition        1.3.0
face-recognition-models 0.3.0
numpy                   1.20.3
opencv-python           4.5.4.58
Pillow                  8.4.0
pip                     21.2.3
pytz                    2021.3
setuptools              57.4.0
sqlparse                0.4.2
wheel                   0.36.2